Updating to remove "non-technical"
Part of fix for https://github.com/ethereum/ethereum-org-website/issues/73

I did not remove the term in the opening paragraph where the text states "This page includes technical **and** non-technical articles". In that context, it is providing the reader useful information about the scope of information on this page and is not using the term in in a diminutive fashion.